trial wrapper for tracking code

+++ wrapTest/factorial.h
@@ -1,2 +1,2 @@
-
-long int factorial(int X);
+#include <stdio.h>
+int factorial(int X);
